+++ scripts/macros/monsterFeatures/construct/shieldGuardian/bound.js
@@ -1,0 +1,4 @@
+export let shieldGuardianBound = {
+    name: 'Bound',
+    version: '1.0.35'
+};